#0: Split topk_utils in cpp and hpp files
