+++ ttnn/cpp/ttnn/operations/reduction/CMakeLists.txt
@@ -45,6 +45,7 @@ target_sources(
         prod/device/prod_op_all.cpp
         prod/prod.cpp
         topk/device/topk_op.cpp
+        topk/device/topk_utils.cpp
         topk/topk.cpp
         topk/device/topk_program_factory.cpp
         reduction_common/reduction_common.cpp
